refactor(#155 PR 9/10): per-type service handlers — strategy pattern registry
Service.js: 1066 -> 756 lines. The per-type if-chain in update()'s job
loop is replaced by a registry dispatch (src/sim/handlers/index.js).
Adding a new service type is now one handler file + one registry line —
the enabler for the Wave 1 feature work (#193).

- 11 handlers: compute, cache, sqs, apigw, cdn, serverless (delegates
  to compute + per-request charge), replica, search, db, nosql, s3.
  waf/alb use the shared genericForward (WAF's malicious blocking lives
  in processQueue intake, not the job chain — unchanged).
- Control-flow contract documented in index.js: handlers return 'next' /
  'requeue-next' / 'requeue-stop', mapped back to the loop's exact
  continue / splice+continue / splice+break semantics.
- Degradation, upkeep, SQS pull, processQueue, ring colors untouched.

This PR is NOT byte-verbatim (control flow was lifted), so equivalence
is proven behaviorally instead: a seeded-RNG (mulberry32) A/B harness
ran 6 scenarios (~1500 steps each) covering every handler incl. SQS
backpressure on HEAD vs this branch — requestsProcessed, failures by
type, money, score, reputation, in-flight all byte-identical in every
scenario.

Also verified: lint 0, 87/87 tests, 42-module graph links, and a live
9-service full-mesh sim (100 spawned -> 96 processed + 4 malicious
blocked, zero failures, zero leaks, zero console errors).
